requirements.txt
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
bcrypt==4.0.1
beautifulsoup4==4.13.5
blinker==1.9.0
blis==1.3.0
boto3==1.28.62
botocore==1.31.85
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.22.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.7
cycler==0.12.1
cymem==2.0.11
deep-translator==1.11.4
Deprecated==1.2.18
dnspython==2.8.0
docx2txt==0.9
email-validator==2.0.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
faiss-cpu==1.12.0
filelock==3.19.1
Flask==2.3.3
flask-babel==4.0.0
flask-cors==6.0.1
Flask-Limiter==3.5.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.7.0
gast==0.4.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
googletrans==4.0.0rc1
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.62.3
gunicorn==21.2.0
h11==0.9.0
h2==3.2.0
h5py==3.14.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.34.4
hyperframe==5.2.0
idna==2.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
keras==3.11.3
kiwisolver==1.4.9
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
libclang==18.1.1
limits==5.6.0
lxml==6.0.1
marisa-trie==1.3.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.6.4
murmurhash==1.0.13
namex==0.1.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
openai==0.28.1
opencv-python==4.12.0.88
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.17.0
ordered-set==4.1.0
packaging==25.0
pandas==2.3.3
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pluggy==1.6.0
preshed==3.0.10
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.8.0
pyparsing==3.0.9
pypdf==6.0.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
reportlab==4.4.3
requests==2.31.0
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==13.9.4
rsa==4.9.1
s3transfer==0.7.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==5.1.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.3
SQLAlchemy==2.0.43
srsly==2.5.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_keras==0.1
termcolor==3.1.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.4
tomli==2.2.1
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.55.4
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.0.7
wasabi==1.1.3
weasel==0.4.1
Werkzeug==2.3.7
wrapt==1.14.2
xgboost==3.0.4
yarl==1.20.1
